requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sentence-transformers
numpy
google-generativeai